requirements.txt
# =============================================================================
# Healthcare Claims RAG Engine — Requirements
# Python 3.11+  |  PEP 8 compliant
# =============================================================================

# --- Web Framework ---
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# --- Data Validation (Pydantic v2) ---
pydantic>=2.7.0
pydantic-settings>=2.3.0

# --- RAG Framework (LlamaIndex 0.11+) ---
llama-index-core>=0.11.0
llama-index-readers-file>=0.2.0
llama-index-embeddings-openai>=0.2.0
llama-index-llms-anthropic>=0.4.0
llama-index-vector-stores-chroma>=0.2.0
llama-index-vector-stores-qdrant>=0.3.0

# --- Vector Stores ---
chromadb>=0.5.0          # Default: zero-friction local store
qdrant-client>=1.10.0    # Optional: swap in via .env VECTOR_STORE_TYPE=qdrant

# --- PDF Parsing ---
pypdf>=4.2.0

# --- LLM & Embeddings ---
anthropic>=0.30.0        # Claude API (decision engine + appeal letters)
openai>=1.30.0           # OpenAI embeddings (text-embedding-3-small)

# --- HTTP & Async ---
httpx>=0.27.0
anyio>=4.4.0

# --- Utilities ---
python-dotenv>=1.0.0
tenacity>=8.3.0          # Retry logic for API calls
structlog>=24.2.0        # Structured logging

# --- Batch Processing & Reporting ---
sqlalchemy>=2.0.0        # SQL ingestion (batch_processor + ehr_connector)
reportlab>=4.0.0         # PDF report generation (batch_processor)
apscheduler>=3.10.0      # Embedded nightly scheduler (nightly_pipeline)

# --- Development & Linting ---
ruff>=0.4.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-httpx>=0.30.0
